+++ src/Custom/Batch/BatchClient.Protocol.cs
@@ -1,6 +1,7 @@
 using System;
 using System.ClientModel;
 using System.ClientModel.Primitives;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace OpenAI.Batch;
@@ -49,10 +50,10 @@ public partial class BatchClient
     /// <param name="options"> The request options, which can override default behaviors of the client pipeline on a per-call basis. </param>
     /// <exception cref="ClientResultException"> Service returned a non-success status code. </exception>
     /// <returns> The response returned from the service. </returns>
-    public virtual async Task<ClientResult> GetBatchesAsync(string after, int? limit, RequestOptions options)
+    public virtual IAsyncEnumerable<ClientResult> GetBatchesAsync(string after, int? limit, RequestOptions options)
     {
-        using PipelineMessage message = CreateGetBatchesRequest(after, limit, options);
-        return ClientResult.FromResponse(await _pipeline.ProcessMessageAsync(message, options).ConfigureAwait(false));
+        BatchesPageEnumerator enumerator = new BatchesPageEnumerator(_pipeline, _endpoint, after, limit, options);
+        return PageCollectionHelpers.CreateAsync(enumerator);
     }
 
     /// <summary>
@@ -63,10 +64,10 @@ public partial class BatchClient
     /// <param name="options"> The request options, which can override default behaviors of the client pipeline on a per-call basis. </param>
     /// <exception cref="ClientResultException"> Service returned a non-success status code. </exception>
     /// <returns> The response returned from the service. </returns>
-    public virtual ClientResult GetBatches(string after, int? limit, RequestOptions options)
+    public virtual IEnumerable<ClientResult> GetBatches(string after, int? limit, RequestOptions options)
     {
-        using PipelineMessage message = CreateGetBatchesRequest(after, limit, options);
-        return ClientResult.FromResponse(_pipeline.ProcessMessage(message, options));
+        BatchesPageEnumerator enumerator = new BatchesPageEnumerator(_pipeline, _endpoint, after, limit, options);
+        return PageCollectionHelpers.Create(enumerator);
     }
 
     /// <summary>