+++ tests/Batch/BatchTests.cs
@@ -4,6 +4,7 @@ using OpenAI.Files;
 using OpenAI.Tests.Utility;
 using System;
 using System.ClientModel;
+using System.Collections.Generic;
 using System.IO;
 using System.Text.Json;
 using System.Threading.Tasks;
@@ -22,32 +23,71 @@ public partial class BatchTests : SyncAsyncTestBase
     }
 
     [Test]
-    public async Task ListBatchesProtocol()
+    public void ListBatchesProtocol()
     {
         BatchClient client = GetTestClient();
-        ClientResult result = IsAsync
-            ? await client.GetBatchesAsync(after: null, limit: null, options: null)
-            : client.GetBatches(after: null, limit: null, options: null);
+        IEnumerable<ClientResult> pageResults = client.GetBatches(after: null, limit: null, options: null);
 
-        BinaryData response = result.GetRawResponse().Content;
-        JsonDocument jsonDocument = JsonDocument.Parse(response);
-        JsonElement dataElement = jsonDocument.RootElement.GetProperty("data");
-
-        Assert.That(dataElement.GetArrayLength(), Is.GreaterThan(0));
-
-        long unixTime2024 = (new DateTimeOffset(2024, 01, 01, 0, 0, 0, TimeSpan.Zero)).ToUnixTimeSeconds();
-
-        foreach (JsonElement batchElement in dataElement.EnumerateArray())
+        int pageCount = 0;
+        foreach (ClientResult pageResult in pageResults)
         {
-            JsonElement createdAtElement = batchElement.GetProperty("created_at");
-            long createdAt = createdAtElement.GetInt64();
+            BinaryData response = pageResult.GetRawResponse().Content;
+            using JsonDocument jsonDocument = JsonDocument.Parse(response);
+            JsonElement dataElement = jsonDocument.RootElement.GetProperty("data");
 
-            Assert.That(createdAt, Is.GreaterThan(unixTime2024));
+            Assert.That(dataElement.GetArrayLength(), Is.GreaterThan(0));
+
+            long unixTime2024 = (new DateTimeOffset(2024, 01, 01, 0, 0, 0, TimeSpan.Zero)).ToUnixTimeSeconds();
+
+            foreach (JsonElement batchElement in dataElement.EnumerateArray())
+            {
+                JsonElement createdAtElement = batchElement.GetProperty("created_at");
+                long createdAt = createdAtElement.GetInt64();
+
+                Assert.That(createdAt, Is.GreaterThan(unixTime2024));
+            }
+            pageCount++;
+
+            //var dynamicResult = result.GetRawResponse().Content.ToDynamicFromJson();
+            //Assert.That(dynamicResult.data.Count, Is.GreaterThan(0));
+            //Assert.That(dynamicResult.data[0].createdAt, Is.GreaterThan(new DateTimeOffset(2024, 01, 01, 0, 0, 0, TimeSpan.Zero)));
         }
 
-        //var dynamicResult = result.GetRawResponse().Content.ToDynamicFromJson();
-        //Assert.That(dynamicResult.data.Count, Is.GreaterThan(0));
-        //Assert.That(dynamicResult.data[0].createdAt, Is.GreaterThan(new DateTimeOffset(2024, 01, 01, 0, 0, 0, TimeSpan.Zero)));
+        Assert.GreaterOrEqual(pageCount, 1);
+    }
+
+    [Test]
+    public async Task ListBatchesProtocolAsync()
+    {
+        BatchClient client = GetTestClient();
+        IAsyncEnumerable<ClientResult> pageResults = client.GetBatchesAsync(after: null, limit: null, options: null);
+
+        int pageCount = 0;
+        await foreach (ClientResult pageResult in pageResults)
+        {
+            BinaryData response = pageResult.GetRawResponse().Content;
+            using JsonDocument jsonDocument = JsonDocument.Parse(response);
+            JsonElement dataElement = jsonDocument.RootElement.GetProperty("data");
+
+            Assert.That(dataElement.GetArrayLength(), Is.GreaterThan(0));
+
+            long unixTime2024 = (new DateTimeOffset(2024, 01, 01, 0, 0, 0, TimeSpan.Zero)).ToUnixTimeSeconds();
+
+            foreach (JsonElement batchElement in dataElement.EnumerateArray())
+            {
+                JsonElement createdAtElement = batchElement.GetProperty("created_at");
+                long createdAt = createdAtElement.GetInt64();
+
+                Assert.That(createdAt, Is.GreaterThan(unixTime2024));
+            }
+            pageCount++;
+
+            //var dynamicResult = result.GetRawResponse().Content.ToDynamicFromJson();
+            //Assert.That(dynamicResult.data.Count, Is.GreaterThan(0));
+            //Assert.That(dynamicResult.data[0].createdAt, Is.GreaterThan(new DateTimeOffset(2024, 01, 01, 0, 0, 0, TimeSpan.Zero)));
+        }
+
+        Assert.GreaterOrEqual(pageCount, 1);
     }
 
     [Test]